refactor: split useTabStore.test.ts into focused test files (#213)
Extract markTerminated/markHostTerminated tests to
useTabStore.terminated.test.ts and persist migration tests to
useTabStore.migration.test.ts. Main file reduced from 517 to 343 lines.
